Move markdown input and select widgets to a widget folder
